rewrite: typescript support
