+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "target": "ES6",
+        "esModuleInterop": true,
+        "module": "commonjs",
+        "declaration": true,
+        "outDir": "./dist",
+        "strict": true 
+    },
+    "include": ["src"],
+    "exclude": ["node_modules", "**/__tests__/**"]
+}
